+++ src/palutils.cpp
@@ -167,7 +167,7 @@ std::string SubstEnvironment(std::string instr)
 	res = "";
 
 	state = 0;
-	len = instr.length();
+	len = (int)instr.length();
 	for (i = 0; i < len; i++)
 	{
 		ch = instr[i];
